@@ -44,6 +44,7 @@
 		"@types/webpack-manifest-plugin": "^2.0.0",
 		"@typescript-eslint/eslint-plugin": "^4.22.0",
 		"@typescript-eslint/parser": "^4.22.0",
+		"@vue/compiler-sfc": "^3.2.47",
 		"autoprefixer": "^9.1.1",
 		"cross-env": "^7.0.3",
 		"css-loader": "^6.7.1",
@@ -71,9 +72,7 @@
 		"tsconfig-paths": "^3.7.0",
 		"typescript": "^4.1.3",
 		"vue-eslint-parser": "^7.1.0",
-		"vue-loader": "^15.9.6",
-		"vue-template-compiler": "^2.6.12",
-		"vue-typescript-import-dts": "^4.0.0",
+		"vue-loader": "^16.0.0",
 		"webpack": "^5.75.0",
 		"webpack-assets-manifest": "^5.0.5",
 		"webpack-cli": "^5.0.1",
@@ -90,9 +89,11 @@
 		"@sentry/browser": "^7.29.0",
 		"@sentry/tracing": "^7.29.0",
 		"@types/orderedmap": "^1.0.0",
+		"@vue/compat": "^3.2.47",
+		"@vue/runtime-dom": "^3.2.47",
 		"axios": "^0.21.4",
 		"bootstrap": "^5.0.0",
-		"bootstrap-vue": "^2.0.0-rc.11",
+		"bootstrap-vue": "^2.23.1",
 		"chart.js": "^2.7.3",
 		"corejs-typeahead": "^1.2.1",
 		"datatables.net": "^1.11.3",
@@ -149,7 +150,7 @@
 		"three": "^0.117.0",
 		"timeago.js": "^4.0.2",
 		"v-money": "^0.8.1",
-		"vue": "^2.6.12"
+		"vue": "^3.1.0"
 	},
 	"lint-staged": {
 		"*": "yarn dlx -q prettier --write --ignore-unknown --plugin=@prettier/plugin-php"
