@@ -1,27 +1,34 @@
-import BootstrapVue from 'bootstrap-vue';
-import Vue from 'vue';
-import Icon from './components/common/Icon.vue';
+import type { App, Component } from 'vue';
+import { createApp } from 'vue';
 import Declaratie from './components/declaratie/Declaratie.vue';
 import Groep from './components/groep/Groep.vue';
 import NamenLeren from './components/namen-leren/NamenLeren.vue';
 import Peiling from './components/peilingen/Peiling.vue';
-import PeilingOptie from './components/peilingen/PeilingOptie.vue';
-import GroepPrompt from './components/editor/GroepPrompt.vue';
 import Inputmask from 'inputmask';
 import money from 'v-money';
 
-Vue.component('icon', Icon);
-Vue.component('peiling', Peiling);
-Vue.component('peilingoptie', PeilingOptie);
-Vue.component('groep', Groep);
-Vue.component('namenleren', NamenLeren);
-Vue.component('declaratie', Declaratie);
-Vue.component('groepprompt', GroepPrompt);
+// Map naam naar vue component
+const vueMap = {
+	peiling: Peiling,
+	groep: Groep,
+	namenleren: NamenLeren,
+	declaratie: Declaratie,
+};
 
-Vue.directive('input-mask', {
-	bind: function (el) {
-		new Inputmask().mask(el);
-	},
-});
-Vue.use(money, { precision: 2, decimal: ',', thousands: ' ', prefix: '€ ' });
-Vue.use(BootstrapVue);
+export const getVueComponent = (naam: string): Component => vueMap[naam];
+
+export const createDefaultApp = (
+	rootComponent: Component,
+	rootProps?: Record<string, unknown>
+): App<Element> => {
+	const app = createApp(rootComponent, rootProps);
+
+	app.directive('input-mask', {
+		beforeMount: function (el) {
+			new Inputmask().mask(el);
+		},
+	});
+	app.use(money, { precision: 2, decimal: ',', thousands: ' ', prefix: '€ ' });
+
+	return app;
+};
